f Move write extension traits to io crate
Got tired on the plane and stopped half way through this.
